backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
SQLAlchemy==2.0.35
aiosqlite==0.20.0
pydantic-settings==2.4.0
opencv-python==4.10.0.84
Pillow==10.4.0
numpy==2.1.1
requests==2.32.3
python-multipart==0.0.9
apify-client==2.1.0
python-dotenv==1.0.0
